Show all demos on home page
