+++ usellm.org/app/demo/prompt-template/page.tsx
@@ -2,7 +2,7 @@
 import { useState } from "react";
 import useLLM from "usellm";
 
-export default function Demo2Page() {
+export default function PromptTemplate() {
   const llm = useLLM({ serviceUrl: "/api/llm" });
   const [topic, setTopic] = useState("");
   const [result, setResult] = useState("");
