+++ usellm.org/app/layout.tsx
@@ -68,7 +68,7 @@ export default function RootLayout({ children }: RootLayoutProps) {
         <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
           <div className="relative flex min-h-screen flex-col">
             <SiteHeader />
-            <div className="flex-1">{children}</div>
+            <div className="flex-1 container">{children}</div>
             <Toaster />
           </div>
           <TailwindIndicator />
