+++ usellm.org/components/examples-nav.tsx
@@ -1,45 +1,15 @@
 "use client";
 
 import Link from "next/link";
-import { usePathname, useSearchParams } from "next/navigation";
+import { useSearchParams } from "next/navigation";
 
 import { cn } from "@/lib/utils";
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
-import { AIChatBotDemo } from "./demos/ai-chatbot-demo";
-import DocumentQnaDemo from "./demos/document-qna-demo";
-
-export const HOME_PAGE_DEMOS = [
-  {
-    name: "AI Chatbot",
-    id: "ai-chatbot",
-    component: AIChatBotDemo,
-  },
-  {
-    name: "Document Q&A",
-    id: "document-qna",
-    component: DocumentQnaDemo,
-  },
-  {
-    name: "Prompt Template",
-    id: "prompt-template",
-    component: AIChatBotDemo,
-  },
-  {
-    name: "Voice Chat",
-    id: "voice-chat",
-    component: AIChatBotDemo,
-  },
-  {
-    name: "Image Generation",
-    id: "image-generation",
-    component: AIChatBotDemo,
-  },
-];
+import { HOME_PAGE_DEMOS } from "@/config/demos";
 
 interface ExamplesNavProps extends React.HTMLAttributes<HTMLDivElement> {}
 
 export function ExamplesNav({ className, ...props }: ExamplesNavProps) {
-  const pathname = usePathname();
   const searchParams = useSearchParams();
   const page = searchParams.get("demo") || "ai-chatbot";
 
@@ -66,13 +36,4 @@ export function ExamplesNav({ className, ...props }: ExamplesNavProps) {
       </ScrollArea>
     </div>
   );
-}
-
-export function HomePageDemo() {
-  const searchParams = useSearchParams();
-  const page = searchParams.get("demo") || "ai-chatbot";
-
-  const C = HOME_PAGE_DEMOS.find((example) => example.id === page)?.component;
-
-  return C ? <C /> : null;
 }
